+++ packages/components-react/src/pages/index.tsx
@@ -52,6 +52,7 @@ import { SegmentedControlPage } from './SegmentedControl';
 import { SelectWrapperPage } from './SelectWrapper';
 import { SelectWrapperSkeletonPage } from './SelectWrapperSkeleton';
 import { SpinnerPage } from './Spinner';
+import { StepperHorizontalPage } from './StepperHorizontal';
 import { SwitchPage } from './Switch';
 import { TabsPage } from './Tabs';
 import { TabsBarPage } from './TabsBar';
@@ -278,6 +279,11 @@ export const generatedRoutes: RouteType[] = [
     name: 'Spinner',
     path: '/spinner',
     element: <SpinnerPage />,
+  },
+  {
+    name: 'Stepper Horizontal',
+    path: '/stepper-horizontal',
+    element: <StepperHorizontalPage />,
   },
   {
     name: 'Switch',