+++ packages/components/src/components/navigation/tabs-bar/tabs-bar-utils.spec.ts
@@ -3,7 +3,6 @@ import {
   determineEnableTransitionClass,
   getFocusedTabIndex,
   getPrevNextTabIndex,
-  getScrollActivePosition,
   getTransformationToActive,
   getTransformationToInactive,
   removeEnableTransitionClass,
@@ -200,50 +199,6 @@ describe('setBarStyle()', () => {
     expect(spy).toHaveBeenCalledWith(0, 1, barElement);
     expect(barElement.classList.contains(enableTransitionClass)).toBe(true);
     expect(barElement.style.cssText).toBe('transform: translate3d(0rem,0,0); width: 0.9375rem;');
-  });
-});
-
-describe('getScrollActivePosition()', () => {
-  it('should return scrollActivePosition = 16 if scrolling to last tab', () => {
-    expect(
-      getScrollActivePosition([{ offsetLeft: 20, offsetWidth: 0 }] as HTMLElement[], 'next', 0, undefined, undefined)
-    ).toBe(16);
-  });
-
-  it('should return scrollActivePosition = 8 if direction is "next", next tab is set as active', () => {
-    expect(
-      getScrollActivePosition(
-        [
-          { offsetLeft: 20, offsetWidth: 0 },
-          { offsetLeft: 0, offsetWidth: 0 },
-        ] as HTMLElement[],
-        'next',
-        0,
-        undefined,
-        20
-      )
-    ).toBe(8);
-  });
-
-  it('should return scrollActivePosition = 0 if direction is "prev" and first tab is set as active', () => {
-    expect(
-      getScrollActivePosition([{ offsetLeft: 0, offsetWidth: 0 }] as HTMLElement[], 'prev', 0, undefined, undefined)
-    ).toBe(0);
-  });
-
-  it('should return scrollActivePosition = 41 if scrolling to previous tab', () => {
-    expect(
-      getScrollActivePosition(
-        [
-          { offsetLeft: 0, offsetWidth: 0 },
-          { offsetLeft: 20, offsetWidth: 5 },
-        ] as HTMLElement[],
-        'prev',
-        1,
-        4,
-        20
-      )
-    ).toBe(41);
   });
 });