+++ packages/components/src/components/navigation/tabs-bar/tabs-bar.tsx
@@ -4,23 +4,18 @@ import {
   attachComponentCss,
   getHTMLElements,
   getPrefixedTagNames,
-  getShadowRootHTMLElement,
+  getScrollActivePosition,
   isShadowRootParentOfKind,
   observeChildren,
   setAttribute,
   unobserveChildren,
 } from '../../../utils';
 import type { TabChangeEvent, TabGradientColorTheme, TabSize, TabWeight } from './tabs-bar-utils';
-import {
-  getFocusedTabIndex,
-  getPrevNextTabIndex,
-  getScrollActivePosition,
-  sanitizeActiveTabIndex,
-  setBarStyle,
-} from './tabs-bar-utils';
+import { getFocusedTabIndex, getPrevNextTabIndex, sanitizeActiveTabIndex, setBarStyle } from './tabs-bar-utils';
 import { getComponentCss } from './tabs-bar-styles';
 import type { BreakpointCustomizable, ThemeExtendedElectric } from '../../../types';
-import type { Direction, ScrollToPosition } from '../../common/scroller/scroller-utils';
+import type { Direction } from '../../common/scroller/scroller-utils';
+import { getScrollerElements } from '../../common/scroller/scroller-utils';
 
 @Component({
   tag: 'p-tabs-bar',
@@ -48,7 +43,6 @@ export class TabsBar {
   @Event({ bubbles: false }) public tabChange: EventEmitter<TabChangeEvent>;
 
   @State() private tabElements: HTMLElement[] = [];
-  @State() private scroll: ScrollToPosition;
 
   private intersectionObserver: IntersectionObserver;
   private barElement: HTMLElement;
@@ -121,7 +115,6 @@ export class TabsBar {
         theme={this.theme}
         gradientColorScheme={this.gradientColorScheme}
         ref={(el) => (this.scrollerElement = el)}
-        scrollToPosition={this.scroll}
         prevNextButtonJssStyle={{
           top: 'calc(50% - .5em)',
           transform: 'translate3d(0,calc(-50% + .375em),0)',
@@ -150,8 +143,9 @@ export class TabsBar {
   };
 
   private defineHTMLElements = (): void => {
-    this.scrollAreaElement = getShadowRootHTMLElement(this.scrollerElement, '.scroll-area');
-    this.prevGradientElement = getShadowRootHTMLElement(this.scrollerElement, '.gradient');
+    const { scrollAreaElement, prevGradientElement } = getScrollerElements(this.scrollerElement);
+    this.scrollAreaElement = scrollAreaElement;
+    this.prevGradientElement = prevGradientElement;
   };
 
   private setTabElements = (): void => {
@@ -227,8 +221,12 @@ export class TabsBar {
       this.prevGradientElement.offsetWidth
     );
 
-    this.scroll = { scrollPosition: scrollActivePosition, isSmooth };
+    (this.scrollerElement as HTMLPScrollerElement).scrollToPosition = {
+      scrollPosition: scrollActivePosition,
+      isSmooth,
+    };
   };
+
   private setBarStyle = (): void => {
     setBarStyle(this.tabElements, this.activeTabIndex, this.barElement, this.prevActiveTabIndex);
   };