+++ packages/storefront/storefront.config.ts
@@ -317,6 +317,18 @@ export const config: StorefrontConfig = {
       // @ts-ignore
       Props: [() => import('@/../../components/src/components/feedback/spinner/spinner.props.md')],
     },
+    'Stepper Horizontal': {
+      // @ts-ignore
+      // prettier-ignore
+      Examples: [() => import('@/../../components/src/components/navigation/stepper-horizontal/stepper-horizontal/stepper-horizontal.examples.md'),],
+      // @ts-ignore
+      // prettier-ignore
+      Usage: [() => import('@/../../components/src/components/navigation/stepper-horizontal/stepper-horizontal/stepper-horizontal.usage.md'),],
+      // @ts-ignore
+      // prettier-ignore
+      Props: [() => import('@/../../components/src/components/navigation/stepper-horizontal/stepper-horizontal/stepper-horizontal.props.md'), () => import('@/../../components/src/components/navigation/stepper-horizontal/stepper-horizontal-item/stepper-horizontal-item.props.md'),
+      ],
+    },
     Switch: {
       // @ts-ignore
       Examples: [() => import('@/../../components/src/components/action/switch/switch.examples.md')],