+++ packages/components/CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+### [2.13.0-rc.4] - 2022-06-22
+
+#### Added
+
+- `Stepper Horizontal`
+
 ### [2.13.0-rc.3] - 2022-06-22
 
 #### Added